# Template n-qbit gate methods.