### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
      If you've fixed a bug or added code that should be tested, add a test to the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and documented.
      If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.

------------------------------------------------------------------------------------------------------------

**Context:**
The class `StateVectorKokkos` has methods to apply special gates (e.g. PauliX) which can be expressed in terms of a templated method.

**Description of the Change:**
Add templated method `applySpecialQubitGate`.
Unify functor interfaces adding a sometimes unused `params` input.
 
**Benefits:**
Reduce boilerplate. 

**Possible Drawbacks:**
Longer compilation time.

**Related GitHub Issues:**