@vincentmr 2023-02-21
> Could you please share the information about compilation time change? It would be great if you can share the system/hardware information that you tests on as well.

It appears the compilation time remains within the margin of error.

# Timings

## Laptop (11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz)

```shell
cmake -Bbuild -DKokkos_ENABLE_OPENMP=On -G Ninja
```

### Old code

```shell
time cmake --build ./build -j2
[32/32] Linking CXX shared module lightning_kokkos_qubit_ops.cpython-38-x86_64-linux-gnu.so

real	1m53.690s
user	3m7.026s
sys	0m5.893s
```

### New code

```shell
time cmake --build ./build -j2
[32/32] Linking CXX shared module lightning_kokkos_qubit_ops.cpython-38-x86_64-linux-gnu.so

real	2m0.505s
user	3m13.847s
sys	0m6.088s
```

## CI

### Old code

[ubuntu-latest (Python 3.11 SERIAL)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/actions/runs/4185498166/jobs/7252678635)
4m 2s

[ubuntu-latest (Python 3.11 OPENMP)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/actions/runs/4185498166/jobs/7252678850)
3m 34s

### New code

[ubuntu-latest (Python 3.11 SERIAL)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/actions/runs/4208141509/jobs/7303811358)
3m 51s

[ubuntu-latest (Python 3.11 OPENMP)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/actions/runs/4208141509/jobs/7303811452)
4m 43s